% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\docType{class}
\name{Association-class}
\alias{Association}
\alias{Association-class}
\title{A Reference Class to return the association object for a gene-disease pair (Ensembl gene ID - EFO ID).}
\description{
A Reference Class to return the association object for a gene-disease pair (Ensembl gene ID - EFO ID).
}
\section{Fields}{

\describe{
\item{\code{utilsObj}}{An instance of class RestUtils}

\item{\code{ensemblGeneID}}{The ensembl gene identifer for the gene of interest (ENSG..).}

\item{\code{efoID}}{An EFO disease name identifier (EFO_...) for the disease of interest.}

\item{\code{associationObjectAsList}}{A list representing the gene-disease association object.}
}}
\section{Methods}{

\describe{
\item{\code{getAssociationObjectAsList()}}{Return the field 'associationObjectAsList'.}

\item{\code{getDataTypesMaxScores()}}{Return a vector with the data types as names for the maximum score values
in the field 'associationObjectAsList'.}

\item{\code{initialize(ensemblGeneID, efoID)}}{Initialize with an Ensembl gene ID and an EFO disease identifier.
Calls an internal method to set the disease-gene list.}

\item{\code{setAssociationObjectAsList()}}{Use the Ensembl gene and EFO disease identifer fields set on initialization to
retrieve a disease-gene association object from the REST API that is used to set the
'associationObjectAsList' field.}
}}

