% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastats.R
\docType{class}
\name{DataStats-class}
\alias{DataStats}
\alias{DataStats-class}
\title{A Reference Class to return information on data types, data sources and counts from the REST utils/stats end-point.}
\description{
A Reference Class to return information on data types, data sources and counts from the REST utils/stats end-point.
}
\section{Fields}{

\describe{
\item{\code{utilsObj}}{An instance of class RestUtils}

\item{\code{allStatsAsList}}{A list that is set in the constructor that contains all the information from the utils/stats end-point.}

\item{\code{dataTypeNames}}{A vector containing all the names of all the data types currently in the Open Targets platform.}

\item{\code{dataSourceNamesForDataTypesList}}{A list mapping data source names to their data type names.}
}}
\section{Methods}{

\describe{
\item{\code{getAllStatsAsList()}}{Return list field value set in in initialize() for all stats for the current version of Open Targets.}

\item{\code{getDataSourceAssocCountsList()}}{Return a list mapping data source names to their associated association counts.}

\item{\code{getDataSourceCountsList(countTypeName)}}{Return a list mapping data source names to either their association or
evidence string counts. The count returned depends on the argument value
which is expected to be either 'associations' or 'evidencestrings'.
Used internally by methods that specify the type of count to return}

\item{\code{getDataSourceEvStrCountsList()}}{Return a list mapping data source names to their associated evidence string counts.}

\item{\code{getDataSourceNamesForDataTypesList()}}{Return field 'dataSourceNamesForDataTypesList' set by method 'setDataSourceNamesForDataTypesList()'}

\item{\code{getDataTypeAssocCountList()}}{Return a list mapping data type names to their association counts.}

\item{\code{getDataTypeCounts(countTypeName)}}{Internal method. Used to get either counts for evidence strings or associations.
Expects a string argument with value of 'evidencestrings' or 'assocations'.
Returns a list mapping data type names to counts for the given argument.}

\item{\code{getDataTypeEvStrCountList()}}{Return a list mapping data type names to their evidence string counts.}

\item{\code{getDataTypeListByName(dataTypeName)}}{Return a list with statistics for a given data type name.}

\item{\code{getDataTypeNames()}}{Return vector field value set in in initialize() for data type names for the current version of Open Targets.}

\item{\code{initialize()}}{Initialize with no arguments.
Calls self methods to set field values.}

\item{\code{setAllStatsAsList()}}{Called in initialize() to set field allStatsAsList.}

\item{\code{setDataSourceNamesForDataTypesList()}}{Internal method.
Return a list of vectors mapping data type names to their associated data sources.
Set return value to a field name by calling in 'initialize()'.}

\item{\code{setDataTypeNames()}}{Internal method. Return data type names as vector.
Called in initialize() to set field dataSourceNames.
Uses the allStatsAsList field set in initialize().}
}}

