% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatypeevidencedataframe.R
\docType{class}
\name{DataTypeEvidenceDataFrame-class}
\alias{DataTypeEvidenceDataFrame}
\alias{DataTypeEvidenceDataFrame-class}
\title{A Reference Class to generate evidence data frames by extracting certain field values from evidence objects.}
\description{
A Reference Class to generate evidence data frames by extracting certain field values from evidence objects.
}
\section{Fields}{

\describe{
\item{\code{evStrChunks}}{A list of lists of evidence objects as returned by method 'getEvStrAll' in class 'EvidenceForDataType'.}
}}
\section{Methods}{

\describe{
\item{\code{getCommonFieldsAsDataFrame()}}{Process the field list 'evStrChunks' and return all the data fields common to all data types as a data frame.}

\item{\code{initialize(evStrChunks)}}{Initialize with a list of lists of ecidence objects.}
}}

