% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\docType{class}
\name{Disease-class}
\alias{Disease}
\alias{Disease-class}
\title{A Reference Class to return information on diseases.}
\description{
A Reference Class to return information on diseases.
}
\section{Fields}{

\describe{
\item{\code{utilsObj}}{An instance of class RestUtils}

\item{\code{diseaseName}}{The disease name as string.}

\item{\code{diseaseNameList}}{A list containing information for a disease.}
}}
\section{Methods}{

\describe{
\item{\code{getAssociationGenesForDiseaseAsDataFrame(diseaseEFO, cutoffScore, direct,
  startFrom = 0)}}{Return a data frame with all association genes for a given disease EFO and set of filters.}

\item{\code{getDiseaseIDNameMap()}}{Return a list mapping EFO IDs to disease names for all diseases returned by the search
engine for the disease term set in 'initialize()'.}

\item{\code{getDiseaseList()}}{Return 'diseaseList' field.}

\item{\code{getEvidenceGenesForDiseaseAsDataFrame(diseaseEFO, cutoffScore, startFrom = 0,
  direct = "true")}}{Return a data frame with all evidence genes for a given disease EFO and set of filters.}

\item{\code{getFirstDiseaseEFOID()}}{Return the EFO for what the search engine regards as the 'best match'.
Check the value returned by the method 'getFirstDiseaseName()'
defined below to ensure that the description matches what is expected.}

\item{\code{getFirstDiseaseName()}}{Return the name for the first disease, assumed to be est match, returned by the search engine.}

\item{\code{getGenesForDiseaseAsDataFrame(filterType, diseaseEFO, startFrom, cutoffScore,
  direct)}}{Used to return both assocoations and evidence genes for a given disease as a data frame.
 The filterType argument determines whether evidences or associations are returned.
 Calls self method 'getGenesForDiseaseAsDataFrameChunk' in a 'repeat' loop to build a data frame
 500 rows at a time. When complete the full data frame is returned.}

\item{\code{getGenesForDiseaseAsDataFrameChunk(filterType, diseaseEFO, startFrom,
  cutoffScore, direct)}}{This method is used internally by the class. It uses the 'filterType' argument to determine whether 'associations' or 'evidence' is returned'.
Return a data frame of gene records for a given EFO ID for 'filterType' =  'evidence/filter?disease' || 'association/filter?disease.}

\item{\code{initialize(diseaseName)}}{Provide a disease name as the single initialization argument.
The underlying search engine can deal with approximate spellings, synonyms and
text case differences.}

\item{\code{setDiseaseList()}}{Set the field 'diseaseList'. Called in initialize().}
}}

