% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence.R
\docType{class}
\name{Evidence-class}
\alias{Evidence}
\alias{Evidence-class}
\title{A Reference Class to implement GET and POST methods to return individual evidence entities as JSON or as R lists.}
\description{
A Reference Class to implement GET and POST methods to return individual evidence entities as JSON or as R lists.
}
\section{Fields}{

\describe{
\item{\code{utilsObj}}{An instance of class RestUtils}
}}
\section{Methods}{

\describe{
\item{\code{getMultipleEvidencesAsJson(evidenceIDList)}}{Convert list returned by method 'getMultipleEvidencesAsList()' to JSON and return the JSON.}

\item{\code{getMultipleEvidencesAsList(evidenceIDList)}}{Given a list of evidence string IDs (32 character hashed values),
return a list of lists where the inner lists are evidence strings.}

\item{\code{getSingleEvidenceAsJson(evidenceID)}}{Convert evidence string for a given ID (32 character hashed value)) to JSON and return the JSON.}

\item{\code{getSingleEvidenceAsList(evidenceID)}}{Use GET to return a list representing the evidence string for a given evidence string ID - a 32 character hashed value.}

\item{\code{initialize()}}{Initialize with no arguments.}

\item{\code{writeEvStrAsJsonToFile(evidenceIDList, fileFullPath)}}{Write JSONs returned for a given list of one or more evidence string IDs to a given file, one JSON per line.}
}}

