% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidencefordatatype.R
\docType{class}
\name{EvidenceForDataType-class}
\alias{EvidenceForDataType}
\alias{EvidenceForDataType-class}
\title{A Reference Class to implement GET and POST methods to return individual evidence entities as JSON or as R lists.}
\description{
A Reference Class to implement GET and POST methods to return individual evidence entities as JSON or as R lists.
}
\section{Fields}{

\describe{
\item{\code{utilsObj}}{An instance of class RestUtils}

\item{\code{dataTypeName}}{The name of a data type in the Open Targets platform.}
}}
\section{Methods}{

\describe{
\item{\code{getEvStrAll()}}{Return a list of lists of all the evidence data for the given data type. Warning: This can be very large.}

\item{\code{getEvStrChunk(chunkSize, startFrom)}}{Return a list of evidence objects of  given chunk size and a given start point.}

\item{\code{initialize(dataTypeName)}}{Initialize with a data type name, e.g. 'genetic_association'.}
}}

