% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidencegenediseasescore.R
\docType{class}
\name{EvidenceGeneDiseaseScore-class}
\alias{EvidenceGeneDiseaseScore}
\alias{EvidenceGeneDiseaseScore-class}
\title{A Reference Class to produce a data frame with all evidence for a given gene, disease and score cut-off.}
\description{
A Reference Class to produce a data frame with all evidence for a given gene, disease and score cut-off.
}
\section{Fields}{

\describe{
\item{\code{utilsObj}}{An instance of class RestUtils}

\item{\code{ensemblGeneID}}{The ensembl gene identifer for the gene of interest (ENSG..).}

\item{\code{efoID}}{An EFO disease name identifier (EFO_...) for the disease of interest.}
}}
\section{Methods}{

\describe{
\item{\code{getAllEvidenceAsDataFrame(directFlag = "false", cutoffScore = 0)}}{Call method 'getDataFramePartial()' repeatedly to generate a complete data frame and return this.}

\item{\code{getDataFramePartial(directFlag, startFrom, chunkSize, cutoffScore = 0)}}{Used internally to return a sub-set of the evidence for the disease and gene field values. Used internally.}

\item{\code{initialize(ensemblGeneID, efoID)}}{Initialize with an ensembl gene ID, an EFO ID.}
}}

