% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\docType{class}
\name{Gene-class}
\alias{Gene}
\alias{Gene-class}
\title{A Reference Class to return information on a given gene.}
\description{
A Reference Class to return information on a given gene.
}
\section{Fields}{

\describe{
\item{\code{utilsObj}}{An instance of class RestUtils}

\item{\code{geneSymbol}}{The gene name or gene symbol as string.}

\item{\code{geneList}}{A list containing platform information for a gene.}
}}
\section{Methods}{

\describe{
\item{\code{getAssociationDiseasesForGeneAsDataFrame(ensemblGeneID, cutoffScore, direct,
  startFrom = 0)}}{Return a data frame with all disease association' for a given ensembl gene ID and set of filters.}

\item{\code{getDiseasesForGeneAsDataFrame(filterType, ensemblGeneID, startFrom, cutoffScore,
  direct)}}{Used to return both associations and evidence genes for a given disease as a data frame.
 The 'filterType' argument determines whether evidences or associations are returned.
Used internally.}

\item{\code{getDiseasesForGeneAsDataFrameChunk(filterType, ensemblGeneID, startFrom,
  cutoffScore, direct)}}{Return a data frame of disease records for a given Ensembl gene ID for 'filterType' =  'evidence' || 'association'.
Used internally.}

\item{\code{getEvidenceDiseasesForGeneAsDataFrame(ensemblGeneID, cutoffScore,
  startFrom = 0, direct = "true")}}{Return a data frame with all evidence diseases for a given ensembl gene ID and set of filters.}

\item{\code{getFirstEnsemblGeneID()}}{Returns the first Ensembl gene ID in 'geneList'.
The first one is assumed to be the best match but needs to be verified.}

\item{\code{getGeneIDNameMap()}}{Return a list mapping Ensembl gene IDs to gene symbols
for all genes returned by the search engine for the gene name
set in 'initialize()'.}

\item{\code{getGeneList()}}{Return field 'geneList'.}

\item{\code{initialize(geneSymbol)}}{Provide a gene symbol or gene name as single argument.
Calls a self method to populate 'geneList'.}

\item{\code{setGeneList()}}{Set field 'geneList'. Called in 'initialize()'.}
}}

