% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restutils.R
\docType{class}
\name{RestUtils-class}
\alias{RestUtils}
\alias{RestUtils-class}
\title{A Reference Class to manage GET and POST calls to the Open Targets REST API}
\arguments{
\item{The}{sub-domain part of the URL to be used by the REST GET API call}
}
\value{
A list summariging the response
}
\description{
A Reference Class to manage GET and POST calls to the Open Targets REST API
}
\section{Fields}{

\describe{
\item{\code{BASE_URL}}{A character string for the REST API URL}
}}
\section{Methods}{

\describe{
\item{\code{checkRESTResponse(response)}}{Check the return code and throw an error if it is not 200.}

\item{\code{getRESTAPIVersion()}}{Return the version of the API being used.}

\item{\code{getRESTPOSTResponse(subdomain, idList)}}{Run a POST request. The identifiers are provided as a standard R list.
Process the response using the method 'processResponse()' and return a list.
Printing the URL used deliberately as a sort of progress indicator and for de-bugging.}

\item{\code{getRESTResponse(subdomain)}}{Run a GET request, process the response and return the resulting list.
Process the response using the method 'processResponse()' and return a list.
Printing the URL used deliberately as a sort of progress indicator and for de-bugging.}

\item{\code{initialize()}}{Initialize with no arguments. Sets the base URL field.
@return .self}

\item{\code{processResponse(response)}}{Used for GET responses.
Returns a three-element list with the status code, the headers and the list returned by the REST call.
@param list A response object as returned by a GET or POST
@return list Three values for status code and response content and header}
}}

